卡券接口

See merge request !64

+++ easy-common/easy-entity/src/main/java/com/rongyi/easy/coupon/entity/mysql/CouponBrand.java
@@ -9,7 +9,7 @@ public class CouponBrand {
 
     private Integer couponId;
 
-    private Byte enabled;
+    private Integer enabled;
 
     public Integer getId() {
         return id;
@@ -43,11 +43,11 @@ public class CouponBrand {
         this.couponId = couponId;
     }
 
-    public Byte getEnabled() {
+    public Integer getEnabled() {
         return enabled;
     }
 
-    public void setEnabled(Byte enabled) {
+    public void setEnabled(Integer enabled) {
         this.enabled = enabled;
     }
 }
