+++ easy-common/easy-entity/src/main/java/com/rongyi/easy/coupon/entity/mysql/CouponMall.java
@@ -13,7 +13,7 @@ public class CouponMall {
 
     private Integer couponId;
 
-    private Byte enabled;
+    private Integer enabled;
 
     public Integer getId() {
         return id;
@@ -63,11 +63,11 @@ public class CouponMall {
         this.couponId = couponId;
     }
 
-    public Byte getEnabled() {
+    public Integer getEnabled() {
         return enabled;
     }
 
-    public void setEnabled(Byte enabled) {
+    public void setEnabled(Integer enabled) {
         this.enabled = enabled;
     }
 }
