+++ easy-common/easy-entity/src/main/java/com/rongyi/easy/coupon/entity/mysql/CouponShop.java
@@ -15,7 +15,7 @@ public class CouponShop {
 
     private Integer couponId;
 
-    private Byte enabled;
+    private Integer enabled;
 
     public Integer getId() {
         return id;
@@ -73,11 +73,11 @@ public class CouponShop {
         this.couponId = couponId;
     }
 
-    public Byte getEnabled() {
+    public Integer getEnabled() {
         return enabled;
     }
 
-    public void setEnabled(Byte enabled) {
+    public void setEnabled(Integer enabled) {
         this.enabled = enabled;
     }
 }
